requirements.txt
Faker
psycopg2-binary
pydantic
pyyaml
python-ulid
bcrypt
ulid
